Extracted BufferBuilder classes
