+++ js/layer_type/fill.js
@@ -13,6 +13,8 @@ module.exports = {
             elementBuffer: 'fillElement',
             secondElementBuffer: 'outlineElement',
 
+            secondElementBufferComponents: 2,
+
             attributes: [{
                 name: 'pos',
                 components: 2,
@@ -34,64 +36,6 @@ module.exports = {
                 }
             }]
 
-        }
-    },
-
-    addFeatures: function() {
-        var features = this.features;
-        for (var i = 0; i < features.length; i++) {
-            var feature = features[i];
-            this.addFeature(feature.loadGeometry());
-        }
-    },
-
-    addFeature: function(lines) {
-        for (var i = 0; i < lines.length; i++) {
-            this.addFill(lines[i]);
-        }
-    },
-
-    addFill: function(vertices) {
-        if (vertices.length < 3) {
-            //console.warn('a fill must have at least three vertices');
-            return;
-        }
-
-        // Calculate the total number of vertices we're going to produce so that we
-        // can resize the buffer beforehand, or detect whether the current line
-        // won't fit into the buffer anymore.
-        // In order to be able to use the vertex buffer for drawing the antialiased
-        // outlines, we separate all polygon vertices with a degenerate (out-of-
-        // viewplane) vertex.
-
-        var len = vertices.length;
-
-        // Check whether this geometry buffer can hold all the required vertices.
-        this.elementGroups.fill.makeRoomFor(len + 1);
-        var elementGroup = this.elementGroups.fill.current;
-
-        var fillVertex = this.buffers.fillVertex;
-
-        // We're generating triangle fans, so we always start with the first coordinate in this polygon.
-        var firstIndex = fillVertex.length - elementGroup.vertexStartIndex,
-            prevIndex, currentIndex, currentVertex;
-
-        for (var i = 0; i < vertices.length; i++) {
-            currentIndex = fillVertex.length - elementGroup.vertexStartIndex;
-            currentVertex = vertices[i];
-
-            this.addFillVertex(currentVertex.x, currentVertex.y);
-
-            // Only add triangles that have distinct vertices.
-            if (i >= 2 && (currentVertex.x !== vertices[0].x || currentVertex.y !== vertices[0].y)) {
-                this.addFillElement(firstIndex, prevIndex, currentIndex);
-            }
-
-            if (i >= 1) {
-                this.addOutlineElement(prevIndex, currentIndex);
-            }
-
-            prevIndex = currentIndex;
         }
     }
